backend/requirements.txt
docker==5.0.3
docker-compose==1.29.2
psycopg2
dockerpty==0.4.1
docopt==0.6.2
#duplicity==0.8.21
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
fpdf==1.7.2
joblib==1.3.2
jsonschema==3.2.0
keyring==23.5.0
#language-selector==0.1
lockfile==0.12.2
#louis==3.20.0
lxml==5.1.0
numpy==1.24.4
oauthlib==3.2.0
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==23.2
paramiko==2.9.3
pdf2image==1.16.3
pexpect==4.8.0
pillow==10.2.0
pytesseract==0.3.10
#python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43
python-docx==1.1.0
python-dotenv==0.19.2
pyttsx3==2.90
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
regex==2023.10.3
requests==2.25.1
sacremoses==0.1.1
safetensors==0.4.1
SecretStorage==3.3.1
sentencepiece==0.1.99
simplejson==3.17.6
six==1.16.0
SQLAlchemy==2.0.23
ssh-import-id==5.11
sympy==1.12
#systemd-python==234
texttable==1.6.4
tokenizers==0.15.0
torch==2.1.2
tqdm==4.66.1
transformers==4.36.2
triton==2.1.0
typing_extensions==4.9.0
argon2-cffi

